# openapi_1.yaml
openapi: 3.0.3
info:
  title: Music & Social
  version: 1.0.4
...
paths:
  /lists:
    post:
      tags:
        - lists
      summary: Erstellen einer Liste
      description: Erstellen einer Liste
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/List"
      responses:
        "201":
          description:
            Liste wurde erstellt
        "400":
          description:
            Liste konnte nicht erstellt werden
        "401":
          description:
            Vorgang nicht möglich Authorisierung notwendig
  /lists/{listId}:
    get:
      tags:
        - lists
      summary: Find list by ID
      description: Returns a single list
      parameters:
        - name: listId
          in: path
          description: DESCRIPTION OF ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: SUCCESSFUL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/List'
        '400':
          description: NOT SUCCESSFUL
        "401":
          description:
            Vorgang nicht möglich Authorisierung notwendig
        "410":
          description:
            Resource steht nicht mehr zur verfügung
    patch:
      tags:
        - lists
      summary: Updaten einer Liste
      parameters:
        - name: listId
          required: true
          in: path
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/List'
      responses:
        '200': 
          description: Liste wurde erfolgreich aktualisiert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/List'
        '400':
          description: Operation unsuccessful
        '404':
          description: Liste existiert nicht
        "401":
          description:
            Vorgang nicht möglich Authorisierung notwendig
    delete:
      tags:
        - lists
      summary: Entfernen einer Liste
      parameters:
        - name: listId
          required: true
          in: path
          schema:
            type: integer
      responses:
        '200': 
          description: Liste wurde erfolgreich entfernt
        '400':
          description: Liste konnte nicht entfernt werden
        '404': 
          description: Liste existiert nicht
        "401":
          description:
            Vorgang nicht möglich Authorisierung notwendig
  /lists/{listId}/items:
    get:
      tags:
        - lists
      summary: Listelemente einer Liste mit Id ausgeben
      description: Gibt ein Array mit Listelementen zurück
      parameters:
        - name: listId
          in: path
          description: Id einer Liste
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: SUCCESSFUL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          description: NOT SUCCESSFUL
        "401":
          description:
            Vorgang nicht möglich Authorisierung notwendig
        "410":
          description:
            Resource steht nicht mehr zur verfügung
...
  /media:
    post:
      tags:
        - media
      summary: Creates new media
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Media'
      responses:
        '201': 
          description: New media was succesfully created
        '400':
          description: Creation was unsuccessful
        "401":
          description:
            Vorgang nicht möglich Authorisierung notwendig
  /media/{mediaId}:
    get:
      tags:
        - media
      summary: Returns media
      parameters:
        - name: mediaId
          required: true
          in: path
          schema:
            type: integer
      responses:
        '200': 
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Media'
        "401":
          description:
            Vorgang nicht möglich Authorisierung notwendig
        "410":
          description:
            Resource steht nicht mehr zur verfügung
    patch: 
      tags:
        - media
      summary: Updates media
      parameters:
        - name: mediaId
          required: true
          in: path
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Media'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Media'
        '400':
          description: Invalid ID supplied
        '404':
          description: Media not found
        "401":
          description:
            Vorgang nicht möglich Authorisierung notwendig
    delete:
      tags:
        - media
      summary: Deletes media
      parameters: 
        - name: mediaId
          required: true
          in: path
          schema:
            type: integer
      responses:
        '200':
          description: Media was successfully deleted
        '400':
          description: Invalid ID value
        "401":
          description:
            Vorgang nicht möglich Authorisierung notwendig
...
components:
  schemas:
    List:
      type: object
      properties: 
        id:
          type: number
        name:
          type: string
        createdAt:
          type: string
        criterias:
          type: array
          items:
            properties:
              key:
                type: string
              value:
                type: string
              operators:
                type: string
    Item:
      type: object
      properties:
        id:
          type: number
        type:
          type: string
        resource:
          type: string
...
    Media:
      type: object
      properties:
        id:
          type: number
        createdAt:
          type: string
          format: date
        path: 
          type: string
...